pyproject.toml
[project]
name = "mlops-churn-prediction"
version = "0.1.0"
description = "MLOps project for real-time customer churn prediction."
requires-python = "~=3.11"
dependencies = [
    "feast[dragonfly]",
    "prometheus-fastapi-instrumentator",
    "fastapi>=0.115.14",
    "uvicorn[standard]>=0.35.0",
    "scikit-learn>=1.7.0",
    "pandas>=2.3.0",
    "mlflow>=3.1.1",
    "apache-airflow>=3.0.2",
    "kafka-python>=2.2.15",
    "dvc>=3.60.1",
    "kagglehub>=0.3.12",
    "black>=25.1.0",
    "plotly>=6.2.0",
    "wandb>=0.21.0",
    "xgboost>=3.0.2",
    "pyyaml>=6.0.2",
    "optuna>=4.4.0",
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.7",
    "nbformat>=5.10.4",
    "seaborn>=0.13.2",
    "requests>=2.32.4",
    "joblib>=1.5.1",
    "ruff>=0.12.1",
    "optuna-integration[wandb]>=4.4.0",
]

[dependency-groups]
dev = [
    "pre-commit>=4.2.0",
    "pytest>=8.4.1",
]
